# maestro-hive/sunday_com_backup_20251005/backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Sunday.com API
  description: |
    Comprehensive API documentation for Sunday.com work management platform.

    ## Authentication
    All API endpoints require authentication via JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - 1000 requests per hour for authenticated users
    - 100 requests per hour for unauthenticated users

    ## Error Responses
    All error responses follow a consistent format:
    ```json
    {
      "error": {
        "type": "validation_error",
        "message": "Invalid request data",
        "details": {},
        "requestId": "uuid"
      }
    }
    ```

    ## Pagination
    List endpoints support cursor and offset-based pagination:
    ```
    GET /api/v1/boards?page=1&limit=20
    ```

  version: 1.0.0
  contact:
    name: Sunday.com Team
    email: api@sunday.com
    url: https://sunday.com/docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sunday.com/terms
servers:
  - url: https://api.sunday.com/api/v1
    description: Production server
  - url: https://staging-api.sunday.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - bearerAuth: []

# ============================================================================
# AUTHENTICATION
# ============================================================================
paths:
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  example: securePassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/register:
    post:
      tags: [Authentication]
      summary: User registration
      description: Create a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: securePassword123
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuthTokens'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: "refresh_token_string"
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

# ============================================================================
# ORGANIZATIONS
# ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        expiresIn:
          type: integer
          description: Access token expiration time in seconds
        user:
          $ref: '#/components/schemas/User'

    # Core Entity Schemas
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          format: uri
          nullable: true
        timezone:
          type: string
          default: "UTC"
        locale:
          type: string
          default: "en"
        settings:
          type: object
          additionalProperties: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message]
          properties:
            type:
              type: string
              example: "validation_error"
            message:
              type: string
              example: "Invalid request data"
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
              format: uuid
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              type: "authentication_error"
              message: "Invalid or expired token"
              requestId: "req_123456"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              type: "validation_error"
              message: "Invalid request data"
              details:
                name: ["Name is required"]
                email: ["Invalid email format"]
              requestId: "req_123456"
...